add cart product in admin page

+++ database/migrations/2024_09_07_084655_create_cart_products_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->unsignedBigInteger('cart_id');
             $table->unsignedBigInteger('product_id');
             $table->unsignedBigInteger('flavor_id');
-            $table->unsignedInteger('stock');
+            $table->bigInteger('stock');
 
             $table->timestamps();
             $table->foreign('cart_id')->references('id')->on('carts')->onDelete('cascade');
